Changed some of the script
M6 was replacing a number with a new number. What was really needed was replacing with the original at all times.
Then, when returning to pos after m6, it would go to a machine coord that was no longer real. i changed the script to record the offset z and return there after m6 ends.